refactoring ccd callback bulkscan controller to separate new file
