+++ src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/callback/CreateBundleIt.java
@@ -75,12 +75,9 @@ public class CreateBundleIt extends AbstractEventIt {
     @Captor
     private ArgumentCaptor<HttpEntity<Callback<SscsCaseData>>> captor;
 
-    @MockitoBean
-    private CcdCallbackHandler ccdCallbackHandler;
-
     @Before
     public void setup() throws IOException {
-        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher, ccdCallbackHandler);
+        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher);
         this.mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
         mapper.findAndRegisterModules();
         json = getJson("callback/createBundleCallback.json");
