+++ src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/callback/EditBundleIt.java
@@ -54,13 +54,10 @@ public class EditBundleIt extends AbstractEventIt {
     @MockitoBean
     private AuthTokenGenerator authTokenGenerator;
 
-    @MockitoBean
-    private CcdCallbackHandler ccdCallbackHandler;
-
 
     @Before
     public void setup() throws IOException {
-        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher, ccdCallbackHandler);
+        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher);
         this.mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
         mapper.findAndRegisterModules();
         json = getJson("callback/editBundleCallback.json");
