+++ src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/callback/IssueFinalDecisionSubmittedHandlerIt.java
@@ -64,12 +64,9 @@ public class IssueFinalDecisionSubmittedHandlerIt extends AbstractEventIt {
     @MockitoBean
     private SscsCcdConvertService sscsCcdConvertService;
 
-    @MockitoBean
-    private CcdCallbackHandler ccdCallbackHandler;
-
     @BeforeEach
     public void setup() throws IOException {
-        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher, ccdCallbackHandler);
+        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher);
         this.mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
         mapper.registerModule(new JavaTimeModule());
         json = getJson("callback/postHearingRequest.json");
