+++ src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/callback/PostHearingReviewIt.java
@@ -65,13 +65,10 @@ public class PostHearingReviewIt extends AbstractEventIt {
 
     @MockitoBean
     private SscsCcdConvertService sscsCcdConvertService;
-
-    @MockitoBean
-    private CcdCallbackHandler ccdCallbackHandler;
         
     @BeforeEach
     public void setup() throws IOException {
-        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher, ccdCallbackHandler);
+        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher);
         this.mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
         mapper.registerModule(new JavaTimeModule());
         json = getJson("callback/postHearingReview.json");
