+++ src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/callback/PostponeHearingRequestIt.java
@@ -62,12 +62,9 @@ public class PostponeHearingRequestIt extends AbstractEventIt {
     @Captor
     private ArgumentCaptor<SscsCaseData> sscsCaseDataArgumentCaptor;
 
-    @MockitoBean
-    private CcdCallbackHandler ccdCallbackHandler;
-
     @BeforeEach
     public void setup() throws IOException {
-        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher, ccdCallbackHandler);
+        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher);
         this.mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
         mapper.registerModule(new JavaTimeModule());
         json = getJson("callback/hearingPostponedRequest.json");
