+++ src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/callback/ReissueFurtherEvidenceIt.java
@@ -38,12 +38,9 @@ public class ReissueFurtherEvidenceIt extends AbstractEventIt {
 
     private String aboutToSubmitPartialJson;
 
-    @MockitoBean
-    private CcdCallbackHandler ccdCallbackHandler;
-
     @Before
     public void setup() throws IOException {
-        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher, ccdCallbackHandler);
+        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher);
         this.mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
         mapper.findAndRegisterModules();
         json = getJson("callback/reissueFurtherEvidenceCallback.json");
