+++ src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/callback/SendToFirstTierIt.java
@@ -65,12 +65,9 @@ public class SendToFirstTierIt extends AbstractEventIt {
     @MockitoBean
     private SscsCcdConvertService sscsCcdConvertService;
 
-    @MockitoBean
-    private CcdCallbackHandler ccdCallbackHandler;
-
     @BeforeEach
     public void setup() throws IOException {
-        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher, ccdCallbackHandler);
+        CcdCallbackController controller = new CcdCallbackController(authorisationService, deserializer, dispatcher);
         this.mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
         mapper.findAndRegisterModules();
         json = getJson("callback/sendToFirstTierRequest.json");
